+++ app/_components/gatheringCard/GatheringHeader.tsx
