+++ app/_components/gatheringModal/GatheringImg.tsx
