+++ app/reviews/data.ts
@@ -1,4 +1,4 @@
-import { Review } from '@/lib/definition';
+import { Review } from '@/types/review';
 
 export const reviews: Review[] = [
   {
