+++ stores/userStore.ts
@@ -1,4 +1,4 @@
-import { User } from '@/lib/definition';
+import { User } from '@/types/user';
 import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
